degroshi/requirements.txt
asgiref==3.7.2
attrs==23.2.0
Django==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
rpds-py==0.17.1
setuptools==69.0.3
sqlparse==0.4.4
swagger-spec-validator==3.0.3

tzdata==2023.4
uritemplate==4.1.1
celery==5.3.6

